requirements.txt
PyYAML==3.10
dropbox==1.5.1
mock==1.0.1
nose==1.3.0
requests==1.2.0
wsgiref==0.1.2
